Refactored tests to remove duplicate code, and fixed Contains/multi-components
